Move and rename tagToObjectRelationApi.ts
Same as tagApi.ts
